+++ .gitignore
@@ -1,3 +1,3 @@
-/node_modules
-/dist
+node_modules
+dist
 .DS_Store
