+++ apps/airview-demo/public/manifest.json
