+++ apps/airview-demo/src/App.js
@@ -1,0 +1,10 @@
+import { greet } from "airview-core";
+
+function App() {
+  const greetJames = greet();
+
+  greetJames("James");
+  return null;
+}
+
+export default App;
