--- demo/src/App.js
@@ -1,5 +1,0 @@
-function App() {
-  return null;
-}
-
-export default App;
