+++ package.json
@@ -2,22 +2,8 @@
   "name": "airview",
   "version": "0.1.0",
   "description": "",
-  "main": "dist/airview.esm.js",
-  "files": [
-    "./lib"
-  ],
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.8.1",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-redux": "^8.0.1",
-    "react-router-dom": "^6.3.0"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/preset-env": "^7.16.11",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-node-resolve": "^13.2.1",
     "eslint": "^8.13.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-jest": "^26.1.4",
@@ -27,16 +13,9 @@
     "eslint-plugin-testing-library": "^5.3.1",
     "husky": "^7.0.4",
     "lint-staged": "^12.4.0",
-    "prettier": "^2.6.2",
-    "rollup": "^2.70.2"
+    "prettier": "^2.6.2"
   },
   "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "lint": "eslint './lib/**/*.{js, jsx}' './demo/src/**/*.{js, jsx}'",
-    "lint:fix": "npm run lint -- --fix",
-    "format": "prettier --check './lib/' './demo'",
-    "format:fix": "npm run format -- --write",
     "prepare": "husky install"
   },
   "repository": {
@@ -50,7 +29,9 @@
   },
   "homepage": "https://github.com/AirWalk-Digital/airview#readme",
   "workspaces": [
-    "demo"
+    "packages/airview-core",
+    "packages/airview-ui",
+    "apps/airview-demo"
   ],
   "lint-staged": {
     "*.js": "eslint",
