+++ packages/airview-core/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "airview-core",
+  "version": "0.1.0",
+  "description": "",
+  "main": "dist/airview.esm.js",
+  "files": [
+    "./dist"
+  ],
+  "peerDependencies": {
+    "@reduxjs/toolkit": "^1.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-redux": "^8.0.1",
+    "react-router-dom": "^6.3.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.9",
+    "@babel/preset-env": "^7.16.11",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-node-resolve": "^13.2.1",
+    "rollup": "^2.70.2"
+  },
+  "scripts": {
+    "build": "rollup -c",
+    "dev": "rollup -c -w",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/AirWalk-Digital/airview.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/AirWalk-Digital/airview/issues"
+  },
+  "homepage": "https://github.com/AirWalk-Digital/airview#readme"
+}
