+++ packages/airview-core/src/features/greet.js
@@ -1,0 +1,7 @@
+export function greet(name) {
+  return () => {
+    const greeting = name ?? "stranger";
+
+    console.log(`Hello ${greeting}!`);
+  };
+}
