+++ packages/airview-ui/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "airview-ui",
+  "version": "0.1.0",
+  "description": "",
+  "main": "index.js",
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/AirWalk-Digital/airview.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/AirWalk-Digital/airview/issues"
+  },
+  "homepage": "https://github.com/AirWalk-Digital/airview#readme"
+}
