--- rollup.config.js
@@ -1,12 +1,0 @@
-import resolve from "@rollup/plugin-node-resolve";
-import babel from "@rollup/plugin-babel";
-import { module } from "./package.json";
-
-export default [
-  {
-    input: "lib/index.js",
-    external: ["react", "react-dom"],
-    output: [{ file: module, format: "es" }],
-    plugins: [resolve(), babel({ babelHelpers: "bundled" })],
-  },
-];
